Research menu
Added research menu comprised of research objects. Research objects can be researched with a button if they are unlocked and you can pay their cost. Researches can be locked or unlocked on start and have a list of research objects they unlock after being researched. If a research object has multiple prerequisite objects, all must be researched before it unlocks them. The menu can be scrolled with a scroll bar.